zapis nowego rpacownika nie diala jeszcze

+++ src/main/java/falynsky/database_project/repository/LocationsRepository.java
@@ -2,7 +2,7 @@ package falynsky.database_project.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface LocationsRepository extends CrudRepository<Locations,Integer> {
+public interface LocationsRepository extends CrudRepository<Locations,Long> {
 
 
 }
