+++ src/main/java/falynsky/database_project/repository/RegionsRepository.java
@@ -2,7 +2,7 @@ package falynsky.database_project.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface RegionsRepository extends CrudRepository<Regions,Integer> {
+public interface RegionsRepository extends CrudRepository<Regions,Long> {
 
 
 }
